scripts/requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
google-genai>=1.0.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.2
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.4.0
numpy>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
